+++ operators/__init__.py
@@ -5,7 +5,6 @@ from .export_uv import ExportUV
 from .export_obj import ExportOBJ, ExportOBJPresetAdd, ExportOBJPresetRemove
 from .export_fbx import ExportFBX, ExportFBXPresetAdd, ExportFBXPresetRemove
 from .collision_maker import CollisionMaker
-from .picking_maker import PickingMaker
 from .clean_setting import CleanSetting
 from .block_rotation_info import BlockRotationInfo
 from .mission_icon_maker import MissionIconMaker
@@ -21,7 +20,6 @@ OPERATOR_CLASSES = (
     ExportFBXPresetAdd,
     ExportFBXPresetRemove,
     CollisionMaker,
-    PickingMaker,
     CleanSetting,
     BlockRotationInfo,
     MissionIconMaker,